refactor(kb-mcp): move build_auth into auth/oidc_proxy.py
auth/__init__.py held the actual OIDC wiring implementation directly,
unlike every tools/*/__init__.py (thin re-export barrel over tool.py +
config.py). Move build_auth/_REQUIRED_SCOPES into oidc_proxy.py and
reduce __init__.py to a re-export, matching that convention.
