+++ services/kb-mcp/tests/test_auth.py
@@ -53,7 +53,7 @@ def test_build_storage_ephemeral_opt_in_returns_a_store(ephemeral_settings):
 
 
 def test_build_auth_never_passes_none_client_storage(durable_settings):
-    with patch("kb_mcp.auth.create_zitadel_oidc_proxy") as mock_create:
+    with patch("kb_mcp.auth.oidc_proxy.create_zitadel_oidc_proxy") as mock_create:
         mock_create.return_value = MagicMock()
         build_auth(durable_settings)
 
@@ -62,7 +62,7 @@ def test_build_auth_never_passes_none_client_storage(durable_settings):
 
 
 def test_build_auth_passes_required_scopes(durable_settings):
-    with patch("kb_mcp.auth.create_zitadel_oidc_proxy") as mock_create:
+    with patch("kb_mcp.auth.oidc_proxy.create_zitadel_oidc_proxy") as mock_create:
         mock_create.return_value = MagicMock()
         build_auth(durable_settings)
 
@@ -82,7 +82,7 @@ def test_build_storage_treats_decryption_errors_as_misses(durable_settings):
 
 
 def test_build_auth_passes_storage_even_in_ephemeral_dev_mode(ephemeral_settings):
-    with patch("kb_mcp.auth.create_zitadel_oidc_proxy") as mock_create:
+    with patch("kb_mcp.auth.oidc_proxy.create_zitadel_oidc_proxy") as mock_create:
         mock_create.return_value = MagicMock()
         build_auth(ephemeral_settings)
